Tighten up kernel documentation with valid notations
The method labeled deprecated is the accepted method used by upstream
AOSP, whereas the alternate is a CAF implementation that was forward
ported from previous configurations. While CM may plan to remove
support, the flexibility of building or not building kernels
independently would be a counterproductive removal. We don't do that.